[ng] Cleanup event display components
