+++ firebird-ng/src/app/main-display/main-display.component.scss
@@ -6,6 +6,4 @@
   padding-right: 0;
 }
 
-#eventDisplay {
-  height: auto;
-}
+
